putting work on the norm
